+++ src/app/shared/directives/highlight/highlight.directive.ts
@@ -1,0 +1,20 @@
+import {Directive, ElementRef, HostBinding, HostListener} from '@angular/core';
+
+@Directive({
+  selector: '[appHighlight]'
+})
+export class HighlightDirective {
+
+  @HostBinding('style.backgroundColor') backgroundColor: string;
+
+  constructor(private el: ElementRef) {
+  }
+
+  @HostListener('mouseenter') onMouseEnter() {
+    this.backgroundColor = 'yellow';
+  }
+
+  @HostListener('mouseleave') onMouseLeave() {
+    this.backgroundColor = null;
+  }
+}