+++ changelog.md
@@ -7,5 +7,20 @@
 - Created "CartService" supports list of added items.
 - Button "Buy" updated to add products to the cart with "CartService".
 - Created "CartComponent" shows list of added items or empty message.
-- Added additional functionality: average ratings.
-- Added additional functionality: cart total.
+- Additional functionality: average ratings.
+- Additional functionality: cart total.
+
+#Task 2:
+- Functionality placed to the few feature modules: "CartModule", "ProductsModule", "SharedModule"
+- Added border around "ProductComponent" items.
+- Added products count to cart.
+- Logic to get count/total extracted to the "CartService".
+- Logic to add products to cart list extracted from "ProductComponent" to the parent.
+- Added new "CartItem" entity and corresponding "CartItemComponent" to show it.
+- Updated "CartService" to maintain list of "CartItem" entities.
+- Added methods to remove items from the card and increase/decrease quantity to the "CartService".
+- Unavailable product displayed with disabled button and grayed text.
+- Project title sets from the "AppComponent" via ViewChild and template variable.
+- OnPush strategy applied to the "ProductsListComponent" and "CartItemComponent".
+- Custom directive created to highlight cart items on mose over.
+- Additional functionality: buttons to remove cart items, increase/decrease quantity.